requirements.txt
mahjong==1.2.1
tweepy==4.14.0
python-dotenv==1.0.1
standard-imghdr==3.13.0
numpy==2.2.1
